Refactor CSS color variables and add light color scheme
The CSS color variables have been changed from old SASS variables to CSS custom properties, improving the compatibility with modern frameworks and tools. The custom properties have been moved inside the :root pseudo-class for global scope. In addition, a light color scheme has been added to the HTML as a part of improving the accessibility of the software.