Move _Layout.cshtml and update styles
The _Layout.cshtml file is relocated from Pages/Shared to Views/Shared. In addition, the `ViewData["title"]` declaration is removed as it was unused. An extra stylesheet (`input.css`) is now included in the head of the document and a new script (`aspnetcore-browser-refresh.js`) is added to the end of the layout file, presumably to support hot-reloads during development.