Update calendar input and cleanup
Modified the handling of the calendar input in calendar.ts and calendar.js to improve UX. The calendar now appropriately adjusts its position based on the input height and adds padding to the top. In addition, the calendar no longer immediately removes on blur but adds a 'hide' class and intended for potential further use. Also removed some unintended random text/comments in the files.