Moved constants from Warehouse.Data to new project Warehouse.Core
The constants that were previously part of Warehouse.Data have been moved to a new project, Warehouse.Core. This change includes renaming the namespace references in the other parts of the solution. A new ApplicationConfiguration class has been introduced in Warehouse.Core as well. The Warehouse.Data project now references this new project.